[refactor] GH-5 Cells renderers in a single file
